Revert changes, simplify child app deployment
