+++ eureka-cli/cmd/deployModules.go
@@ -68,9 +68,9 @@ func DeployModules() {
 
 	slog.Info(deployModulesCommand, internal.GetFuncName(), "### PULLING SIDECAR IMAGE ###")
 	deployModulesDto := internal.NewDeployModulesDto(vaultRootToken, map[string]string{internal.FolioRegistry: "", internal.EurekaRegistry: ""}, registryModules, backendModulesMap, environment, sidecarEnvironment)
-	sidecarImage := internal.GetSidecarImage(deployModulesCommand, deployModulesDto.RegistryModules[internal.EurekaRegistry])
+	sidecarImage := internal.GetSidecarImage(deployManagementCommand, deployModulesDto.RegistryModules[internal.EurekaRegistry])
 	sidecarResources := internal.CreateResources(false, viper.GetStringMap(internal.SidecarModuleResourcesKey))
-	internal.PullModule(deployModulesCommand, client, sidecarImage)
+	internal.PullModule(deployManagementCommand, client, sidecarImage)
 
 	slog.Info(deployModulesCommand, internal.GetFuncName(), "### DEPLOYING MODULES ###")
 	deployedModules := internal.DeployModules(deployModulesCommand, client, deployModulesDto, sidecarImage, sidecarResources)
