+++ eureka-cli/cmd/redirect.go
@@ -38,7 +38,8 @@ var redirectCmd = &cobra.Command{
 // TODO Fix redirectModules to work on host network
 func RedirectModules() {
 	slog.Info(redirectModuleCommand, internal.GetFuncName(), "### REDIRECT MODULE ###")
-	internal.UpdateApplicationModuleDiscovery(redirectModuleCommand, enableDebug, strings.ReplaceAll(id, ":", "-"), location, restore, internal.DefaultServerPort)
+	id = strings.ReplaceAll(id, ":", "-")
+	internal.UpdateApplicationModuleDiscovery(redirectModuleCommand, enableDebug, id, location, restore, internal.DefaultServerPort)
 }
 
 func init() {
